XMR Config class loads pkg config
