+++ Connector/xmr/handler.py
@@ -4,7 +4,7 @@ from httputils import httpmethod
 from rpcutils import rpcmethod, rpcutils, error
 from logger import logger
 from utils import utils
-from .connector import Config
+from .config import Config
 from .constants import COIN_SYMBOL
 
 
@@ -21,19 +21,17 @@ class Handler:
             logger.printError(f"Configuration {network} already added for {self.coin}")
             return False, f"Configuration {network} already added for {self.coin}"
 
-        defaultConfig, err = utils.loadDefaultPackageConf(self.coin)
-        if err is not None:
-            return False, err
-
-        self._networksConfig[network] = Config(
-            networkName=network,
-            config={
-                "protocol": config["protocol"] if "protocol" in config else defaultConfig["protocol"],
-                "host": config["host"] if "host" in config else defaultConfig["host"],
-                "rpcPort": config["rpcPort"] if "rpcPort" in config else defaultConfig["rpcPort"],
-                "rpcPath": config["rpcPath"] if "rpcPath" in config else defaultConfig["rpcPath"]
-            }
+        pkgConfig = Config(
+            coin=self.coin,
+            networkName=network
         )
+
+        ok, err = pkgConfig.loadConfig(config=config)
+        if not ok:
+            logger.printError(f"Can not load config for {network} for {self.coin}: {err}")
+            return ok, err
+
+        self.networksConfig[network] = pkgConfig
 
         return True, None
 
@@ -45,7 +43,7 @@ class Handler:
 
         return self.networksConfig[network].jsonEncode(), None
 
-    def removeConfig(self, network):
+    async def removeConfig(self, network):
 
         if network not in self.networksConfig:
             logger.printError(f"Configuration {network} not added for {self.coin}")
@@ -55,25 +53,16 @@ class Handler:
 
         return True, None
 
-    def updateConfig(self, network, config):
+    async def updateConfig(self, network, config):
 
         if network not in self.networksConfig:
             logger.printError(f"Configuration {network} not added for {self.coin}")
             return False, f"Configuration {network} not added for {self.coin}"
 
-        defaultConfig, err = utils.loadDefaultPackageConf(self.coin)
-        if err is not None:
-            return False, err
-
-        self.networksConfig[network] = Config(
-            networkName=network,
-            config={
-                "protocol": config["protocol"] if "protocol" in config else defaultConfig["protocol"],
-                "host": config["host"] if "host" in config else defaultConfig["host"],
-                "rpcPort": config["rpcPort"] if "rpcPort" in config else defaultConfig["rpcPort"],
-                "wsPort": config["wsPort"] if "wsPort" in config else defaultConfig["wsPort"]
-            }
-        )
+        ok, err = self.networksConfig[network].loadConfig(config=config)
+        if not ok:
+            logger.printError(f"Can not load config for {network} for {self.coin}: {err}")
+            return ok, err
 
         return True, None
